requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
dj-database-url==0.5.0
Django==3.2.9
django-environ==0.4.5
django-heroku==0.3.1
django-storages==1.12.3
docopt==0.6.2
google-api-core==2.7.1
google-auth==2.6.0
google-cloud-core==2.2.3
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
google-site-verification==0.0.4
googleapis-common-protos==1.55.0
gunicorn==20.1.0
idna==3.3
Js2Py==0.71
packaging==21.3
Pillow==9.0.1
pipwin==0.5.1
protobuf==3.19.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyjsparser==2.7.1
pyparsing==3.0.6
PyPrind==2.11.3
pySmartDL==1.3.4
python-decouple==3.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rsa==4.8
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
whitenoise==5.3.0
